+++ app/src/components/Comment/CommentHeader/EditionDate.tsx
@@ -1,17 +1,26 @@
 import React from 'react';
 
-import { Typography } from '@material-ui/core';
+import { makeStyles, Theme, Typography } from '@material-ui/core';
 import dayjs from 'dayjs';
 
 const DATE_FORMAT = '[Le] DD.MM.YYYY [à] HH:mm';
 
+const useStyles = makeStyles<Theme, { canViewHistory: boolean }>(() => ({
+  edited: ({ canViewHistory }) => ({
+    fontStyle: 'italic',
+    cursor: canViewHistory ? 'pointer' : 'inherit',
+  }),
+}));
+
 type EditionDateProps = {
   edited: boolean;
   date: Date;
-  onViewHistory: () => void;
+  onViewHistory?: () => void;
 };
 
 const EditionDate: React.FC<EditionDateProps> = ({ edited, date, onViewHistory }) => {
+  const classes = useStyles({ canViewHistory: Boolean(onViewHistory) });
+
   if (!edited) {
     return (
       <Typography variant="caption">
@@ -24,7 +33,7 @@ const EditionDate: React.FC<EditionDateProps> = ({ edited, date, onViewHistory }
     <Typography
       variant="caption"
       title="Édité"
-      style={{ fontStyle: 'oblique', cursor: 'pointer' }}
+      className={classes.edited}
       onClick={onViewHistory}
     >
       *&nbsp;{ dayjs(date).format(DATE_FORMAT) }